+++ src/strongbox/cli.clj
@@ -7,7 +7,6 @@
    [me.raynes.fs :as fs]
    [strongbox
     [zip :as zip]
-    [constants :as constants]
     [joblib :as joblib]
     [github-api :as github-api]
     [tukui-api :as tukui-api]
@@ -17,7 +16,6 @@
     [logging :as logging]
     [addon :as addon]
     [specs :as sp]
-    [catalogue :as catalogue]
     [http :as http]
     [utils :as utils :refer [if-let* message-list]]
     [core :as core :refer [get-state paths]]]))
@@ -628,72 +626,12 @@
 
 ;; importing addons
 
-;; todo: logic might be better off in core.clj
-(defn-spec find-addon (s/or :ok :addon/summary, :error nil?)
-  "given a URL of a supported addon host, parses it, looks for it in the catalogue, expands addon and attempts to install a dry run installation.
-  if successful, returns the addon-summary.
-  dry-run installation attempt can be skipped by setting `attempt-dry-run` to false."
-  [addon-url string?, attempt-dry-run? boolean?]
-  (binding [http/*cache* (core/cache)]
-    (if-let* [addon-summary-stub (catalogue/parse-user-string addon-url)
-              source (:source addon-summary-stub)
-              match-on-list [[[:source :url] [:source :url]]
-                             [[:source :source-id] [:source :source-id]]]
-              addon-summary (cond
-                              (= source "github")
-                              (or (github-api/find-addon (:source-id addon-summary-stub))
-                                  (error (message-list
-                                          "Failed. URL must be:"
-                                          ["valid"
-                                           "originate from github.com"
-                                           "addon uses 'releases'"
-                                           "latest release has a packaged 'asset'"
-                                           "asset must be a .zip file"
-                                           "zip file must be structured like an addon"])))
-
-                              (= source "gitlab")
-                              (or (gitlab-api/find-addon (:source-id addon-summary-stub))
-                                  (error (message-list
-                                          "Failed. URL must be:"
-                                          ["valid"
-                                           "originate from gitlab.com"
-                                           "addon uses releases"
-                                           "latest release has a custom asset with a 'link'"
-                                           "link type must be either a 'package' or 'other'"])))
-
-                              (= source "curseforge")
-                              (error (str "addon host 'curseforge' was disabled " constants/curseforge-cutoff-label "."))
-
-                              :else
-                              ;; look in the current catalogue. emit an error if we fail
-                              (or (:catalogue-match (core/-find-first-in-db (or (core/get-state :db) []) addon-summary-stub match-on-list))
-                                  (error (format "couldn't find addon in catalogue '%s'"
-                                                 (name (core/get-state :cfg :selected-catalogue))))))
-
-              ;; game track doesn't matter when adding it to the user catalogue.
-              ;; prefer retail though (it's the most common) and `strict` here is `false`
-              addon (or (catalogue/expand-summary addon-summary :retail false)
-                        (error "failed to fetch details of addon"))
-
-              ;; a dry-run is good when importing an addon for the first time but
-              ;; not necessary when updating the user-catalogue.
-              _ (if-not attempt-dry-run?
-                  true
-                  (or (core/install-addon-guard addon (core/selected-addon-dir) true)
-                      (error "failed dry-run installation")))]
-
-             ;; if-let* was successful!
-             addon-summary
-
-             ;; failed if-let* :(
-             nil)))
-
 (defn-spec import-addon nil?
   "goes looking for given `addon-url` and, if found, adds it to the user catalogue and then installs it."
   [addon-url string?]
   (binding [http/*cache* (core/cache)]
     (if-let* [dry-run? true
-              addon-summary (find-addon addon-url dry-run?)
+              addon-summary (core/find-addon addon-url dry-run?)
               addon (core/expand-summary-wrapper addon-summary)]
              ;; success! add to user-catalogue and proceed to install
              (do (core/add-user-addon! addon-summary)
@@ -706,41 +644,9 @@
              ;; failed to find or expand summary, probably because of selected game track.
              nil)))
 
-(defn-spec refresh-user-catalogue-item nil?
-  "refresh the details of an individual `addon` in the user catalogue, optionally writing the updated catalogue to file."
-  [addon :addon/summary, db :addon/summary-list]
-  (logging/with-addon addon
-    (info "refreshing user-catalogue entry")
-    (try
-      (let [{:keys [source source-id url]} addon
-            refreshed-addon (core/db-addon-by-source-and-source-id db source source-id)
-            attempt-dry-run? false
-            refreshed-addon (or refreshed-addon
-                                (find-addon url attempt-dry-run?))]
-        (if-not refreshed-addon
-          (warn "failed to refresh user-catalogue entry as the addon was not found in the catalogue or online")
-          (core/add-user-addon! refreshed-addon)))
-
-      (catch Exception e
-        (error (format "an unexpected error happened while refreshing the user-catalogue entry: %s" (.getMessage e)))))))
-
-(defn-spec refresh-user-catalogue nil?
-  "refresh the details of all addons in the user catalogue, writing the updated catalogue to file once."
+(defn refresh-user-catalogue
   []
-  (binding [http/*cache* (core/cache)]
-    (let [path (fs/base-name (core/paths :user-catalogue-file)) ;; "user-catalogue.json"
-          ;; we can't assume the full-catalogue is available.
-          _ (core/download-catalogue (core/get-catalogue-location :full))
-          db (catalogue/read-catalogue (core/catalogue-local-path :full))
-          full-catalogue (or (:addon-summary-list db) [])]
-      (info (format "refreshing \"%s\", this may take a minute ..." path))
-      (doseq [user-addon (core/get-state :user-catalogue :addon-summary-list)]
-        (refresh-user-catalogue-item user-addon full-catalogue))
-      (core/write-user-catalogue!)
-      (info (format "\"%s\" has been refreshed" path))))
-  nil)
-
-;;
+  (core/refresh-user-catalogue))
 
 ;; todo: shift to core.clj or addon.clj
 (defn-spec addon-source-map-to-url (s/or :ok ::sp/url, :error nil?)