+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* an (opt-in) automatic refesh of the user-catalogue every 28 days.
+    - see `Preferences` -> `Keep user catalogue updated`.
+    - the user-catalogue is a catalogue of the addons added to strongbox using `File` -> `Import addon` 
+        - or through 'starring' a regular catalogue addon.
 * a "clear" button to the search addons tab that removes all search filters, including search terms.
 * new column for installed addons "starred" that will add an installed addon to the 'user-catalogue'.
     - star button disabled when addon is being ignored or isn't matched against the catalogue.